refactoring screen containers and changing file locations
These changes do not affect functionality for users.
